refactor: 日付・申請状態バッジヘルパーを app 直下の共有場所へ昇格
見積申請一覧画面（#572）が消費者に加わり、日付ヘルパ（fromDateInputValue/
toDateInputValue）とバッジ tone 写像（badgeToneOf/badgeToneClassName）の
消費者が2画面になる。単一ソース化の適期として estimates 配下から app 直下の
共有場所（_lib/date.ts・_components/shared/variationApplicationStateBadge.ts）
へ昇格し、estimates 詳細側の import を張り替える。挙動不変（既存テスト green）。
