+++ src/app/(features)/estimates/[estimateNumber]/EstimateHeaderForm.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { useServerForm } from "@/app/_hooks/useServerForm";
 import { SelectionModal } from "@/app/_components/shared/SelectionModal";
 import type { SearchFieldDef } from "@/app/_components/shared/SearchForm";
-import { toDateInputValue } from "../_shared/date";
+import { toDateInputValue } from "@/app/_lib/date";
 import { inputClassDisabled } from "../_shared/formStyles";
 import { FkSelectionField } from "../_shared/FkSelectionField";
 import { RepairDetailFields } from "../_shared/RepairDetailFields";
