+++ src/app/(features)/estimates/[estimateNumber]/page.tsx
@@ -7,7 +7,7 @@ import {
   resolveEffectiveTaxRateQueryFactory,
 } from "@subdomains/estimate/application/factories/estimateQueryFactory";
 import { notFound } from "next/navigation";
-import { fromDateInputValue, toDateInputValue } from "../_shared/date";
+import { fromDateInputValue, toDateInputValue } from "@/app/_lib/date";
 import { DuplicateEstimateModal } from "./DuplicateEstimateModal";
 import { EstimateHeaderSection } from "./EstimateHeaderSection";
 import { VariationPanel } from "./VariationPanel";
