+++ src/app/(features)/estimates/new/actions.ts
@@ -8,7 +8,7 @@ import { checkTaxRateThenCreateDepsFactory } from "@subdomains/estimate/applicat
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { handleCommandError } from "../../_shared/error-handler";
-import { fromDateInputValue } from "../_shared/date";
+import { fromDateInputValue } from "@/app/_lib/date";
 import { taxRateMismatchFormErrors } from "../_shared/tax-rate-format";
 import { toVariationContentInputFromNodes } from "../[estimateNumber]/variationContentMapping";
 import { createEstimateSchema } from "./schema";
