+++ src/app/_components/shared/variationApplicationStateBadge.ts
